#229 Upgraded django and lots of dependencies. Django-reversion upgrade required some work on a signal in GlossVideo. Django-summernotes migration was finally fixed. #247 still remains open, as there has not been any upgrades to django-registration, and I think the problem with confusable-homoglyphs still exists.

+++ signbank/video/models.py
@@ -153,7 +153,7 @@ class GlossVideo(models.Model):
 @receiver(models.signals.pre_delete, sender=GlossVideo)
 def delete_on_delete(sender, instance, **kwargs):
     """On signal delete, use pre_delete to delete the videofile. This will happen when a GlossVideo object is deleted"""
-    if instance.videofile:
+    if isinstance(instance, GlossVideo) and instance.videofile:
         if os.path.isfile(instance.videofile.path):
             os.remove(instance.videofile.path)
